+++ src/types/api/index.ts
@@ -1,3 +1,4 @@
+import type { AudioMethod } from "./audio";
 import type { CatalogMethod } from "./catalog";
 
-export type VKAudioMethod = CatalogMethod;
+export type VKAudioMethod = CatalogMethod | AudioMethod;