+++ src/types/client/audio.ts
@@ -1,0 +1,6 @@
+import type { AudioItem } from "./section";
+
+export type SearchAudioResult = {
+  count: number;
+  audios: AudioItem[];
+};